requirements.txt
fastapi
uvicorn
mysql-connector-python
python-dotenv
jinja2
firebase_admin
pydantic
bcrypt
PyJWT
python-multipart
cryptography
pytz
websockets
wsproto
boto3
redis